requirements.txt
# --- 深度学习核心 ---
torch
torchvision
pytorch_lightning
einops

# --- 数据处理与工具 ---
numpy
tqdm
pyyaml
pillow

# --- [新增] 数据集下载 ---
# train_model_zoo.py 中使用了 requests.get 下载 TinyImageNet
requests

# --- [新增] 可视化与交互环境 ---
# 用于运行 Visualization.ipynb, operator.ipynb 等
matplotlib
jupyter